Fixed Win32 warnings

+++ tests/test_test.c
@@ -73,8 +73,8 @@ INA_TEST(test, assert_data) {
 }
 
 INA_TEST(test, assert_equal_floating) {
-    const float exp = 3.3;
-    const float real = 3.3;
+    const double exp = 3.3;
+    const double real = 3.3;
     INA_TEST_ASSERT_EQUAL_FLOATING(exp, real);
 }
